[bdk_chain_redesign] MOVE: `IndexedTxGraph` into submodule
